+++ Arbiter.App/Models/DebugSettings.cs
@@ -9,10 +9,14 @@ public class DebugSettings : ICloneable
     public bool ShowMonsterId { get; set; }
     public bool ShowMonsterClickId { get; set; }
     public bool ShowHiddenPlayers { get; set; }
+    public bool ShowPlayerNames { get; set; }
+    public bool DisableWeatherEffects { get; set; }
+    public bool EnableTabMap { get; set; }
     public bool IgnoreEmptyMessages { get; set; }
 
     public bool CheckEnabled() => ShowDialogId || ShowNpcId || ShowMonsterId || ShowMonsterClickId ||
-                                  ShowHiddenPlayers || IgnoreEmptyMessages;
+                                  ShowHiddenPlayers || ShowPlayerNames || EnableTabMap || DisableWeatherEffects ||
+                                  IgnoreEmptyMessages;
 
     public object Clone() => new DebugSettings
     {
@@ -21,6 +25,9 @@ public class DebugSettings : ICloneable
         ShowMonsterId = ShowMonsterId,
         ShowMonsterClickId = ShowMonsterClickId,
         ShowHiddenPlayers = ShowHiddenPlayers,
+        ShowPlayerNames = ShowPlayerNames,
+        EnableTabMap = EnableTabMap,
+        DisableWeatherEffects = DisableWeatherEffects,
         IgnoreEmptyMessages = IgnoreEmptyMessages
     };
 }